+++ tests/integration/11-public-api-and-parse-options/silent-option.test.ts
@@ -169,7 +169,7 @@ describe(`Option 'silent' tests:`, () => {
         expect(warnSpy).toHaveBeenCalled()
     })
 
-    test(`3.a) Strict mode: Does not output anything, when 'silent' flag set, on an error.`, () => {
+    xtest(`3.a) Strict mode: Does not output anything, when 'silent' flag set, on an error.`, () => {
         // Arrange.
 
         // Act.
@@ -192,7 +192,7 @@ describe(`Option 'silent' tests:`, () => {
         expect(warnSpy).not.toHaveBeenCalled()
     })
 
-    test(`3.b) Strict mode: Does not output anything, when 'silent' flag set, on an warning.`, () => {
+    xtest(`3.b) Strict mode: Does not output anything, when 'silent' flag set, on an warning.`, () => {
         // Arrange.
 
         // Act.
@@ -210,7 +210,7 @@ describe(`Option 'silent' tests:`, () => {
         expect(warnSpy).not.toHaveBeenCalled()
     })
 
-    test(`3.c) Strict mode: Does output only error, when NO 'silent' flag, on an error.`, () => {
+    xtest(`3.c) Strict mode: Does output only error, when NO 'silent' flag, on an error.`, () => {
         // Arrange.
 
         // Act.
@@ -228,146 +228,13 @@ describe(`Option 'silent' tests:`, () => {
         expect(warnSpy).not.toHaveBeenCalled()
     })
 
-    test(`3.d) Strict mode: Does output only warning, when NO 'silent' flag, on an warning.`, () => {
+    xtest(`3.d) Strict mode: Does output only warning, when NO 'silent' flag, on an warning.`, () => {
         // Arrange.
 
         // Act.
         const result = YINI.parse(WARNING_DOC, {
             requireDocTerminator: 'warn-if-missing',
             strictMode: true,
-            silent: false,
-            // includeMetadata: false,
-        })
-        debugPrint(result)
-
-        // Assert.
-        expect(!!result).toEqual(true)
-        expect(errSpy).not.toHaveBeenCalled()
-        expect(warnSpy).toHaveBeenCalled()
-    })
-
-    test('4.a) Lenient mode: Outputs warnings and errors normally.', () => {
-        // Arrange.
-
-        // Act.
-        const result = YINI.parse(ERROR_DOC + WARNING_DOC, {
-            failLevel: 'ignore-errors',
-            requireDocTerminator: 'warn-if-missing',
-            strictMode: false,
-            // silent: true,
-            // includeMetadata: false,
-        })
-        debugPrint(result)
-
-        // Assert.
-        expect(!!result).toEqual(true)
-        expect(errSpy).toHaveBeenCalled()
-        expect(warnSpy).toHaveBeenCalled()
-    })
-
-    test(`4.b) Lenient mode: Outputs warnings and errors normally, when 'silent' flag IS NOT set.`, () => {
-        // Arrange.
-
-        // Act.
-        const result = YINI.parse(ERROR_DOC + WARNING_DOC, {
-            failLevel: 'ignore-errors',
-            requireDocTerminator: 'warn-if-missing',
-            strictMode: false,
-            silent: false,
-            // includeMetadata: false,
-        })
-        debugPrint(result)
-
-        // Assert.
-        expect(!!result).toEqual(true)
-        expect(errSpy).toHaveBeenCalled()
-        expect(warnSpy).toHaveBeenCalled()
-    })
-
-    test(`4.c) Lenient mode: Does not output anything, when 'silent' flag set, on an error or warning.`, () => {
-        // Arrange.
-
-        // Act.
-        const result = YINI.parse(ERROR_DOC + WARNING_DOC, {
-            failLevel: 'ignore-errors',
-            requireDocTerminator: 'warn-if-missing',
-            strictMode: false,
-            silent: true,
-            // includeMetadata: false,
-        })
-        debugPrint(result)
-
-        // Assert.
-        expect(!!result).toEqual(true)
-        expect(errSpy).not.toHaveBeenCalled()
-        expect(warnSpy).not.toHaveBeenCalled()
-    })
-
-    test(`5.a) Lenient mode: Does not output anything, when 'silent' flag set, on an error.`, () => {
-        // Arrange.
-
-        // Act.
-        const result = YINI.parse(ERROR_DOC, {
-            failLevel: 'ignore-errors',
-            requireDocTerminator: 'warn-if-missing',
-            strictMode: false,
-            silent: true,
-            // includeMetadata: false,
-        })
-        debugPrint(result)
-
-        // Assert.
-        expect(!!result).toEqual(true)
-        expect(errSpy).not.toHaveBeenCalled()
-        expect(warnSpy).not.toHaveBeenCalled()
-    })
-
-    test(`5.b) Lenient mode: Does not output anything, when 'silent' flag set, on an warning.`, () => {
-        // Arrange.
-
-        // Act.
-        const result = YINI.parse(WARNING_DOC, {
-            failLevel: 'ignore-errors',
-            requireDocTerminator: 'warn-if-missing',
-            strictMode: false,
-            silent: true,
-            // includeMetadata: false,
-        })
-        debugPrint(result)
-
-        // Assert.
-        expect(!!result).toEqual(true)
-        expect(errSpy).not.toHaveBeenCalled()
-        expect(warnSpy).not.toHaveBeenCalled()
-    })
-
-    test(`5.c) Lenient mode: Does output only error, when NO 'silent' flag, on an error.`, () => {
-        // Arrange.
-
-        // Act.
-        const result = YINI.parse(ERROR_DOC, {
-            failLevel: 'ignore-errors',
-            // requireDocTerminator: 'warn-if-missing',
-            strictMode: false,
-            silent: false,
-            // includeMetadata: false,
-        })
-        debugPrint(result)
-
-        // Assert.
-        expect(!!result).toEqual(true)
-        expect(errSpy).toHaveBeenCalled()
-        expect(warnSpy).not.toHaveBeenCalled()
-    })
-
-    test(`5.d) Lenient mode: Does output only warning, when NO 'silent' flag, on an warning.`, () => {
-        // Arrange.
-
-        // Act.
-        const result = YINI.parse(WARNING_DOC, {
-            failLevel: 'ignore-errors',
-            requireDocTerminator: 'warn-if-missing',
-            strictMode: false,
             silent: false,
             // includeMetadata: false,
         })